new version that uses Personal Access Token and a LED Strip (nodemcu)
To improve security, now you can use a Personal Access Token instead of a user/password combination.

Also this version currenlty not uses Bluetooth light bulbs, but an LED strip, that is controlled through a custom program running on a NodeMCU (EPS8266). The code for that project will be provided later.